v1.0.4 release
* Improved the README
* Touched up a few nonconformist docstrings in math_utils
* Fixed some typos in the math_utils.limit() docstring
* Added math_utils.factor() and math_utils.ifactor(), which factor a given number, and added appropriate tests to test_libs.py
* Removed an over-optimization and added a small optimization to math_utils.factors() (see notes below)
* Added some testing code for math_utils.factors()

Notes on math_utils.factors() optimization changes:
The root-plus-2 optimizer on the range was too large; taken from a prime tester, it left out numbers for (for instance) all even numbers over 10.
It was replaced by the half-plus-1 optimizer, which does not have this problem.
The code for making a set for @n was improved slightly by exchanging the enclosing list for a tuple. This improvement was found to be more than 5 times faster, though the time consumed in the first place was minuscule.